requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
google-genai
pydantic
python-dotenv